Add custom hook for login logic and modularize login component
