+++ components/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './AuthLogo';
+export * from './FormLogin';
+export * from './ButtonProvider';
