+++ hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useProducts';
+export * from './useAuthForm';
